Backend

See merge request s09-webmobile1-sub2/S09P12B111!83

+++ Backend/Chu/src/main/java/com/chu/consulting/domain/Review.java
@@ -10,14 +10,14 @@ import javax.persistence.Embeddable;
 @Getter @Setter @ToString
 public class Review {
 
-    private Float reviewScore;
+    private Double reviewScore;
     private String reviewContent;
 
     protected Review() {
 
     }
 
-    public Review(Float reviewScore, String reviewContent) {
+    public Review(Double reviewScore, String reviewContent) {
         this.reviewScore = reviewScore;
         this.reviewContent = reviewContent;
     }
